refactor(budget): TransactionDirection-Enum einführen und Domain-API bereinigen
- TransactionType-Enum aus BudgetTransactionEntity extrahiert → eigenständige
  TransactionDirection-Klasse im Domain-Package
- BudgetTransactionEntity: Konstruktor vereinfacht (yearMonth wird automatisch
  aus bookingDate abgeleitet via setBookingDate()), alle Signaturen auf
  TransactionDirection umgestellt
- BudgetRepository: findDefaultActiveAccountId() und applyExpenseToAccountBalance()
  hinzugefügt, getEffectiveLimitCents() entfernt, Signaturen auf TransactionDirection
  umgestellt
- ImportStatus in das Domain-Importing-Package verschoben, ClaudeApiKeyStore
  nach data/api verschoben
- AppDatabase: unzulässige Migration-Konstanten entfernt, nur
  fallbackToDestructiveMigration() behalten
- TemplateStatusUpdate als neuen Domain-Typ hinzugefügt
- CLAUDE.md gekürzt und präzisiert
- Alle Aufrufer (DAOs, Repositories, Services, UI) entsprechend angepasst

+++ src/main/java/com/autosecretary/features/budget/domain/internal/recurring/SuggestionScorer.java
@@ -48,9 +48,9 @@ public final class SuggestionScorer {
 
     public static double calculateConfidence(List<RecurringBudgetTransaction> txList,
                                              DatePatternDetector.PatternResult pattern,
-                                             int avgAmount,
-                                             int minAmount,
-                                             int maxAmount) {
+                                             long avgAmount,
+                                             long minAmount,
+                                             long maxAmount) {
         double score = 0;
         score += Math.min(txList.size() / OCCURRENCE_CAP, OCCURRENCE_WEIGHT);
